Move the smoothing code into cleaning_methods to be consistent
Also move the GpsSmoothing code into unused_tests since it is not that relevant
any more.  In particular, it uses data from the old format without accuracies,
so it is hard to test our new code against it. Will need to create new data files here.
